fix: caso o usuario buscado para ser deletado for null

+++ src/main/java/com/kelvin/shareTrip/service/implemetation/UsuarioServiceImplementation.java
@@ -42,7 +42,9 @@ public class UsuarioServiceImplementation implements UsuarioService {
     @Override
     public Usuario deleteUsuario(String id) {
         Usuario usuarioDelete = usuarioRepo.findById(id).orElse(null);
-        usuarioRepo.delete(usuarioDelete);
+        if (usuarioDelete != null) {
+            usuarioRepo.delete(usuarioDelete);
+        }
         return usuarioDelete;
     }
